v2: Delete old version and replace with v2
