+++ pages/basicExample.mdx
@@ -7,7 +7,7 @@ import { Playground, PropsTable } from 'docz';
 import 'tachyons/css/tachyons.css'; // eslint-disable-line
 import Wizard from '../src/Wizard';
 import Step from '../src/Step';
-import Controls from '../src/Controls';
+import Controls from './components/Controls';
 import Emoji from './components/Emoji';
 import Button from './components/Button';
 
@@ -22,41 +22,17 @@ The following is a pretty basic example of how it works. We have a `<Wizard />` 
     <Wizard
       onComplete={() => {
         console.log('wizard completed');
-      }}
-      render={() => (
-        <>
-          <Step name="first">
-            <Emoji className="f-headline lh-copy" emoji="☝️" />
-            <div className="mt2 f6 lh-copy">
-              <p className="mv0">Step 1</p>
-            </div>
-          </Step>
-          <Step name="second">
-            <Emoji className="f-headline lh-copy" emoji="✌️" />
-            <div className="mt2 f6 lh-copy">
-              <p className="mv0">Step 2</p>
-            </div>
-          </Step>
-          <Step name="third">
-            <Emoji className="f-headline lh-copy" emoji="🤟" />
-            <div className="mt2 f6 lh-copy">
-              <p className="mv0">Step 3</p>
-            </div>
-          </Step>
-          <Controls
-            render={(onNext, onPrevious, isFirstStep, isLastStep) => (
-              <div className="flex items-center justify-center pa4">
-                <Button onClick={onPrevious} disabled={isFirstStep}>
-                  <Emoji emoji="👈" /> <span className="ml3">Previous</span>
-                </Button>
-                <Button onClick={onNext} disabled={isLastStep}>
-                  <span className="mr3">Next</span> <Emoji emoji="👉" />
-                </Button>
-              </div>
-            )}
-          />
-        </>
-      )}
-    />
+      }}>
+      <Step name="first">
+        <Emoji className="f-headline lh-copy" emoji="☝️" />
+      </Step>
+      <Step name="second">
+        <Emoji className="f-headline lh-copy" emoji="✌️" />
+      </Step>
+      <Step name="third">
+        <Emoji className="f-headline lh-copy" emoji="🤟" />
+      </Step>
+      <Controls />
+    </Wizard>
   </div>
 </Playground>
