--- src/index.js
@@ -1,5 +1,0 @@
-import Wizard from './Wizard';
-import Step from './Step';
-import Controls from './Controls';
-
-export { Wizard, Step, Controls };
